requirement.txt
streamlit
open ai